tex_dolphin renderMapBlock 83.36->100 (unit 98.94->99.75): case-4 advance loop via ModelRenderInstrsState* sp local - pointer-var access blocks the register promotion that collapsed the unrolled body to +64 (retail keeps per-iteration load/add/store, forwarding only within the unrolled group); named base walker reused in case 4 kills the instrs reload; off anchored by (pos = state.bit) >> 3 side-effecting initializer (plain decl gets propagated to its use below base's load); callList passSelect u32->u8 drops the hoisted clrlwi, callList itself byte-MATCH with the (u8) cast gone; decl order instructions/done/shader/doSetup/instructionCount from permsweep. fnbytes MATCH 612 B, DOL retail e750e8

+++ include/main/tex_dolphin.h
@@ -31,7 +31,7 @@ struct MapShader* mapBlockRender_setShader(u8 doSetup, struct MapBlockData* bloc
 struct MapShader* mapBlockRender_setShader(int doSetup, struct MapBlockData* blockData,
                                            ModelRenderInstrsState* state);
 #endif
-void mapBlockRender_callList(u32 passSelect, u32 visArg, struct MapBlockData* block, struct MapShader* shader,
+void mapBlockRender_callList(u8 passSelect, u32 visArg, struct MapBlockData* block, struct MapShader* shader,
                              ModelRenderInstrsState* state, float* mtx);
 
 #endif /* MAIN_TEX_DOLPHIN_H_ */
